fetch2: Unable to fetch URL should be a warning, not an error
Running a build server with local sstate cache and running the
oe-core sstate-cache-management.sh script on cron can cause errors if
an sstate file is removed before it can be used. This is not an
error because the package can be rebuilt from source. Changing
the message to a warning is more appropriate in this case.

(LOCAL REV: NOT UPSTREAM) -- Sent to bitbake-devel on 20131204

Signed-off-by: Konrad Scherer <Konrad.Scherer@windriver.com>
Signed-off-by: Jeff Polk <jeff.polk@windriver.com>