bitbake: runqueue/cooker: catch hashvalidate error in RunQueueExecuteScenequeue
We need to catch the failure of self.rq.hashvalidate which invoked
in RunQueueExecuteScenequeue, and exit in a normal build. So we
raise bb.BBHandledException for this specific error.

Tweak execute_runqueue's exception process order. So the build will
exit peacefully while the above failure occurs.

(LOCAL REV: NOT UPSTREAM) -- Sent to bitbake-devel on 20140827

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>